update: renaming from cabotage to subql
